+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# <a href='https://github.com/mrjackwills/adsbdb/releases/tag/v0.4.3'>v0.4.3</a>
+### 2025-05-21
+
+### Chores
++ .devcontainer updated, [e18b8d24](https://github.com/mrjackwills/adsbdb/commit/e18b8d2402765a28bf943743bb084c4e7306f265)
++ dependencies updated, [6f62c302](https://github.com/mrjackwills/adsbdb/commit/6f62c302aece6c99990c396b4bef14c963328e63), [83898caf](https://github.com/mrjackwills/adsbdb/commit/83898cafc089cd7b92605ece8e10c83a25b63081)
++ run.sh updated, [6912c600](https://github.com/mrjackwills/adsbdb/commit/6912c600c65d5abc1fa232f5e5cc274d2ef64cfa)
+
+### Fixes
++ postgres migrations & port, [0993fa66](https://github.com/mrjackwills/adsbdb/commit/0993fa66c3970fc23b9c7ec4fb4755b9cb696cde)
+
+### Refactors
++ LazyStatic api version, [79d94493](https://github.com/mrjackwills/adsbdb/commit/79d944936b680e747f74efebdfbc9ef9420767ae)
++ dead code removed, [d13c2922](https://github.com/mrjackwills/adsbdb/commit/d13c292227613d7a091dfc9d71a336fa0798c67d)
++ use ApiRoutes struct, [2fda1483](https://github.com/mrjackwills/adsbdb/commit/2fda148387f4253f33b13a45f291419bf3846bbb)
+
+### Tests
++ tests updated with new flightroute data, [36d47d47](https://github.com/mrjackwills/adsbdb/commit/36d47d479cfc5b7fcbaa4265e04e87bac7c76294), [bb04dd5d](https://github.com/mrjackwills/adsbdb/commit/bb04dd5d4515463a3d73bbe6b6db78316e3f4a69)
+
 # <a href='https://github.com/mrjackwills/adsbdb/releases/tag/v0.4.2'>v0.4.2</a>
 ### 2025-04-11